requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.6.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
yarl==1.9.2
